requirements.txt
# ============================================
# MCP Server Tutorials - 의존성 패키지
# ============================================

# 핵심 패키지
fastmcp>=2.0.0          # FastMCP 프레임워크
python-dotenv>=1.0.0     # 환경변수 관리

# Tutorial 02: 날씨 서버
httpx>=0.27.0            # 비동기 HTTP 클라이언트

# Tutorial 04: MSSQL 데이터베이스 (선택)
# sqlalchemy>=2.0.0      # ORM / DB 연결
# pyodbc>=5.0.0          # MSSQL ODBC 드라이버

# Tutorial 05: 메모앱 (선택)
# aiosqlite>=0.20.0      # 비동기 SQLite
